polish ckeditor; mvp for #28, using browser confirm; close #28

+++ src/components/application-status/application-status-component.tsx
@@ -97,7 +97,7 @@ class ApplicationStatusComponent extends Component<
                     applicationStatus ? (
                         <IconButton
                             onClick={() =>
-                                this.props.deleteApplicationStatus(
+                                confirm(`Are you sure you want to delete this status? ${applicationStatus.text}`) && this.props.deleteApplicationStatus(
                                     applicationStatus
                                 )
                             }
